fix: remove `loadFont` from the store
